+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+ - Exception support with CATCH and THROW.
+ - ABORT and ABORT" throw exceptions -1 and -2.
  - S" now works in interpretation mode. It returns a transient buffer.
- - STRING, stores a counted string.
+ - 0> to compat.
 ### Fixed
  - V: crash when entering ":w!"